edit: move test class to pacakage step1
