--- src/test/kotlin/study/StringKoTest.kt
@@ -1,13 +1,0 @@
-package study
-
-import io.kotest.core.spec.style.AnnotationSpec
-import io.kotest.matchers.shouldBe
-
-class StringKoTest : AnnotationSpec() {
-    @Test
-    fun isEmpty() {
-        "".isEmpty() shouldBe true
-        " ".isEmpty() shouldBe false
-        "a".isEmpty() shouldBe false
-    }
-}
