+++ src/test/kotlin/study/step1/Person.kt
@@ -1,3 +1,3 @@
-package study
+package study.step1
 
 data class Person(val name: String, val age: Int, var nickname: String? = name)
