+++ src/test/kotlin/study/step1/PersonKoTest.kt
@@ -1,4 +1,4 @@
-package study
+package study.step1
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.inspectors.forAll
